Refined File structure
